pyproject.toml
[project]
name = "ai-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.5.0",
    "arxiv>=2.4.0",
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi[standard]>=0.129.0",
    "openai>=2.20.0",
    "pyjwt>=2.11.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.54.0",
    "yfinance>=1.1.0",
]
